Missed adding a class to AttachedSobjEffect
